# pied_piper_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pied Piper",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "get users by search",
                "operationId": "search for users",
                "parameters": [
                    {
                        "name": "searrch",
                        "in": "query",
                        "description": "enter name of user you want to search for",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all found users"
                    },
                    "401": {
                        "description": "unauthorized."
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
        "/api/profile/{id}": {
            "get": {
                "tags": [
                    "user_profile"
                ],
                "summary": "get user profile",
                "operationId": "profile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of your profile or of other user profile",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user profile"
                    },
                    "401": {
                        "description": "unauthorized."
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/updatePicture": {
            "post": {
                "tags": [
                    "user_profile"
                ],
                "summary": "update picture",
                "description": "User update his picutre here",
                "operationId": "update_picture",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "picture": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "picture updated successfully!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/updateCover": {
            "post": {
                "tags": [
                    "user_profile"
                ],
                "summary": "update cover",
                "description": "User update his cover here",
                "operationId": "update_cover",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "cover": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "cover updated successfully!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/userPictures": {
            "get": {
                "tags": [
                    "user_profile"
                ],
                "summary": "get user pictures",
                "operationId": "profile_pictures",
                "responses": {
                    "200": {
                        "description": "user pictures"
                    },
                    "401": {
                        "description": "unauthorized."
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/updateJobTitles": {
            "post": {
                "tags": [
                    "user_profile"
                ],
                "summary": "update JobTitles",
                "description": "User update his JobTitles here",
                "operationId": "update_JobTitles",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "available_to_hire",
                                    "job_title",
                                    "location"
                                ],
                                "properties": {
                                    "available_to_hire": {},
                                    "job_title": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "JobTitles updated successfully!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}